created model views
refactored repositories:
- created ModelViews
- delegated repo logic to components
- repositories are wrappers
